+++ instrumentation/opentelemetry-instrumentation-celery/src/opentelemetry/instrumentation/celery/utils.py
@@ -65,7 +65,7 @@ def set_attributes_from_context(span, context):
             if value in [(None, None), [None, None]]:
                 continue
             if None in value:
-                value = ["" if tl is None else tl for tl in value]
+                value = ["" if tl is None else str(tl) for tl in value]
 
         # Skip `retries` if it's value is `0`
         if key == "retries" and value == 0: